@@ -5,35 +5,36 @@
     "license": "MIT",
     "authors": [
         {
-            "name": "Martin Ruf",
-            "email": "Martin.Ruf@rerp.de"
+            "name": "Marek Baron",
+            "email": "baron.marek@googlemail.com"
         }
     ],
     "require": {
-        "php": "^7.4|^8.0",
-        "psr/container": "^1.0|^2.0"
+        "php": "^8.1",
+        "psr/container": "^2.0"
     },
     "require-dev": {
         "roave/security-advisories": "dev-master",
         "phpunit/phpunit": "^9.5",
-        "phpstan/phpstan": "^1",
+        "phpstan/phpstan": "^1.2",
         "phpstan/phpstan-phpunit": "^1.0",
-        "jangregor/phpstan-prophecy": "^1.0",
-        "phpspec/prophecy-phpunit": "^2.0"
+        "phpspec/prophecy-phpunit": "^2.2"
     },
     "autoload": {
         "psr-4": {
-            "teewurst\\Pipeline\\": "./src"
+            "Baron\\Pipeline\\": "./src"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "teewurst\\Pipeline\\test\\": "./test"
+            "Baron\\Pipeline\\test\\": "./test"
         }
     },
     "scripts": {
         "analyse": "phpstan analyse src -l 9",
-        "test": "phpunit -c phpunit.xml",
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf",
+        "test-unit": "phpunit -c phpunit.xml",
         "coverage": "phpunit -c phpunit.xml --coverage-clover=coverage.xml"
     }
 }
