@@ -1,17 +1,16 @@
 <?php
+
 declare(strict_types=1);
 
-namespace teewurst\Pipeline;
+namespace Baron\Pipeline;
 
 /**
  * Interface PayloadInterface
  *
  * Object with is passed through all tasks and returned from a pipeline
  *
- * @deprecated Will be removed in later versions - Payload should be a term only and used as freely as possible
- *
- * @package teewurst\Pipeline
- * @author  Martin Ruf <Martin.Ruf@check24.de>
+ * @package Baron\Pipeline
+ * @author  Marek Baron<baron.marek@googlemail.com>
  */
 interface PayloadInterface
 {
